Update run_first.py
This change goes alongside the recent change to "requirements.txt" to help resolve issue #4. It makes it so people who have already installed older versions of the libraries can easily update them. It's not the perfect implementation for a fix like this, but it saves some headaches for the time being.